tidy ups
